requirements.txt
django
djangorestframework-simplejwt
djangorestframework
django-cors-headers